Added Map to table
Now the map  is shown inside the visualization table. As a limitant, it only appears in the latitude column. The longitude column still shows the default visualization.
Also, the table now is fixed in size: every column is 200px width. This is because otherwise, the map couldnt be shown inside it. Now I'm trying to add a button to show a bigger visualization under the table.